+++ app/src/components/app.component.ts
@@ -10,6 +10,26 @@ import * as packageJson from "../../../package.json"
 
 const testKey = localStorage?.stripeAnguarKey || "pk_test_5JZuHhxsinNGc5JanVWWKSKq"
 const stripeServer = 'https://api.stripe.com/v1/';
+const sampleAddress = {
+  city: 'Coconut Creek',
+  country: null,
+  line1: '1234 sw 1st ct',
+  line2: null,
+  postal_code: '33066',
+  state: 'FL'
+}
+const sample = {
+  metadata: {
+    testedUsing: 'stripe-angular',
+    author: 'Acker Apple'
+  },
+  owner: {
+    email: 'jenny.rosen@example.com',
+    name: 'jenny rosen',
+    phone: '561-561-5611',
+    address: sampleAddress
+  }
+}
 
 @Component({
   selector:"app",
@@ -34,6 +54,7 @@ const stripeServer = 'https://api.stripe.com/v1/';
   card: {
     token?:any
     source?:any
+    payment_method?: stripe.paymentMethod.PaymentMethod
   } = {}
   stripe:stripe.Stripe
   stripeBank:stripe.Stripe
@@ -62,25 +83,13 @@ const stripeServer = 'https://api.stripe.com/v1/';
   };
 
   // passed along when token or sources created
-  request = {
-    owner: {
-      "address": {
-        "city": 'Coconut Creek',
-        "country": null,
-        "line1": '1234 sw 1st ct',
-        "line2": null,
-        "postal_code": "33066",
-        "state": 'FL'
-      },
-      email: 'jenny.rosen@example.com',
-      name: 'jenny rosen',
-      phone: '561-561-5611',
-    },
+  sourceRequest = {
+    owner: sample.owner,
+    metadata: sample.metadata
+  }
 
-    metadata: {
-      testedUsing: 'stripe-angular',
-      author: 'Acker Apple'
-    }
+  paymentMethodRequest = {
+    metadata: sample.metadata
   }
 
   // passed along during card token creation
@@ -93,7 +102,7 @@ const stripeServer = 'https://api.stripe.com/v1/';
     address_line2: "",
     address_state: "",
     address_zip: "",
-    metadata: this.request.metadata
+    metadata: sample.metadata
   }
 
   // ach token data
@@ -111,7 +120,7 @@ const stripeServer = 'https://api.stripe.com/v1/';
       account_number: '000123456789',
       account_holder_name: 'Jenny Rosen',
       account_holder_type: 'individual',
-      metadata: this.request.metadata
+      metadata: sample.metadata
     } as (stripe.BankAccountTokenOptions) // The stripe-v3 types are missing the metadata property.
   }
 
@@ -119,15 +128,51 @@ const stripeServer = 'https://api.stripe.com/v1/';
     $send: new EventEmitter(),
     data: {
       description: "some new customer",
-      metadata: this.request.metadata
+      ...sample.owner,
+      metadata: sample.metadata
+    }
+  }
+
+  customer_update: ISimpleRouteEditor = {
+    $send: new EventEmitter(),
+    data: {
+      id: "",
+      metadata: sample.metadata
+    }
+  }
+
+  customer_get: ISimpleRouteEditor = {
+    $send: new EventEmitter(),
+    data: {
+      id: ""
     }
   }
 
   source_get: ISimpleRouteEditor = {
     $send: new EventEmitter(),
     data: {
-      description: "get source by id",
-      source: ""
+      id: ""
+    }
+  }
+
+  source_update: ISimpleRouteEditor = {
+    $send: new EventEmitter(),
+    data: {
+      id: ""
+    }
+  }
+
+  payment_method_get: ISimpleRouteEditor = {
+    $send: new EventEmitter(),
+    data: {
+      id: ""
+    }
+  }
+
+  payment_method_update: ISimpleRouteEditor = {
+    $send: new EventEmitter(),
+    data: {
+      id: ""
     }
   }
 
@@ -135,9 +180,10 @@ const stripeServer = 'https://api.stripe.com/v1/';
     $send: new EventEmitter(),
     data: {
       amount: 1099,
+      confirm: 'true',
       currency: 'usd',
       setup_future_usage: 'off_session',
-      metadata: this.request.metadata
+      metadata: sample.metadata
     }
   }
 
@@ -147,12 +193,17 @@ const stripeServer = 'https://api.stripe.com/v1/';
       amount: 1099,
       currency: 'usd',
       source: 'token-here',
-      metadata: this.request.metadata
+      metadata: sample.metadata
     }
   }
 
   constructor(public StripeScriptTag: StripeScriptTag){
-    this.source_get.$send.subscribe(data => this.getSource(data.source))
+    this.source_update.$send.subscribe(data => this.sendSourceUpdate(data, data.id))
+    this.payment_method_update.$send.subscribe(data => this.sendPaymentMethodUpdate(data, data.id))
+    this.payment_method_get.$send.subscribe(data => this.getPaymentMethod(data.id))
+    this.source_get.$send.subscribe(data => this.getSource(data.id))
+    this.customer_get.$send.subscribe(data => this.getCustomer(data.id))
+    this.customer_update.$send.subscribe(data => this.updateCustomer(data, data.id))
     this.customer.$send.subscribe(data => this.createCustomer(data))
     this.payintent.$send.subscribe(data => this.createPayIntent(data))
     this.charge.$send.subscribe(data => this.createCharge(data))
@@ -191,11 +242,19 @@ const stripeServer = 'https://api.stripe.com/v1/';
       .then(stripe =>this.stripe=stripe);
   }
 
-  changeRequest(data:string){
-    this.request = JSON.parse(data)
+  changeSourceRequest(data:string){
+    this.sourceRequest = JSON.parse(data)
 
-    if (this.request.metadata) {
-      this.extraData.metadata = this.request.metadata
+    if (this.sourceRequest.metadata) {
+      this.extraData.metadata = this.sourceRequest.metadata
+    }
+  }
+
+  changePaymentMethodRequest(data:string){
+    this.paymentMethodRequest = JSON.parse(data)
+
+    if (this.paymentMethodRequest.metadata) {
+      this.extraData.metadata = this.paymentMethodRequest.metadata
     }
   }
 
@@ -215,7 +274,13 @@ const stripeServer = 'https://api.stripe.com/v1/';
       current = current[keys.shift()];
     }
 
-    current[keys[0]] = JSON.parse(value);
+    try {
+      // current[keys[0]] = JSON.parse(value);
+      eval('current[keys[0]] = ' + value); // allow loose js to be cast to json
+    } catch (err) {
+      console.error(`failed to parse object key ${keys[0]}`);
+      throw err
+    }
   }
 
   log(message) {
@@ -237,7 +302,34 @@ const stripeServer = 'https://api.stripe.com/v1/';
     request({
       url: stripeServer + 'sources/' + sourceId,
       authorizationBearer: this.privateKey
-    }).then(res => this.source_get.result = tryParse(res));
+    }).then(res => this.setSource(res));
+  }
+
+  getPaymentMethod(id: string) {
+    request({
+      url: stripeServer + 'payment_methods/' + id,
+      authorizationBearer: this.privateKey
+    }).then(res => this.setPaymentMethod(res));
+  }
+
+  setPaymentMethod(res: any) {
+    this.payment_method_get.result = tryParse(res)
+    this.payment_method_get.resultAt = Date.now()
+  }
+
+  setSource(res: any) {
+    this.source_get.result = tryParse(res)
+    this.source_get.resultAt = Date.now()
+  }
+
+  getCustomer(id: string) {
+    request({
+      url: stripeServer + 'customers/' + id,
+      authorizationBearer: this.privateKey
+    }).then(res => {
+      this.customer_get.result = tryParse(res)
+      this.customer_get.resultAt = Date.now()
+    });
   }
 
   createCustomer(data: any) {
@@ -245,7 +337,10 @@ const stripeServer = 'https://api.stripe.com/v1/';
       url: stripeServer + 'customers',
       post: data,
       authorizationBearer: this.privateKey
-    }).then(res => this.customer.result = tryParse(res));
+    }).then(res => {
+      this.customer.result = tryParse(res)
+      this.customer.resultAt = Date.now()
+    });
   }
 
   createPayIntent(data: any) {
@@ -253,7 +348,10 @@ const stripeServer = 'https://api.stripe.com/v1/';
       url: stripeServer + 'payment_intents',
       post: data,
       authorizationBearer: this.privateKey
-    }).then(res => this.payintent.result = tryParse(res));
+    }).then(res => {
+      this.payintent.result = tryParse(res)
+      this.payintent.resultAt = Date.now()
+    });
   }
 
   createCharge(data: any) {
@@ -261,12 +359,24 @@ const stripeServer = 'https://api.stripe.com/v1/';
       url: stripeServer + 'charges',
       post: data,
       authorizationBearer: this.privateKey
-    }).then(res => this.charge.result = tryParse(res));
+    }).then(res => {
+      this.charge.result = tryParse(res)
+      this.charge.resultAt = Date.now()
+    });
   }
 
+  // a source or token converted into a customer
   createCustomerByToken(token: stripe.Token) {
     const customer = this.customer.data;
     customer.source = token.id;
+
+    this.createCustomer(customer);
+  }
+
+  // a source or token converted into a customer
+  createCustomerByPaymentMethod(data: stripe.paymentMethod.PaymentMethod) {
+    const customer = this.customer.data;
+    customer.payment_method = data.id;
 
     this.createCustomer(customer);
   }
@@ -287,8 +397,118 @@ const stripeServer = 'https://api.stripe.com/v1/';
       }
     }).then(result => this.bank.verifyResponse = tryParse(result))
   }
-}
 
+  fetchPayIntentUpdate(config: ISimpleRouteEditor) {
+    const base = stripeServer + 'payment_intents/'
+    const intentId = config.result.id;
+    const url = base + intentId;
+    request({
+      url,
+      authorizationBearer: this.privateKey
+    }).then(result => config.retrieve = tryParse(result))
+  }
+
+  sendSourceUpdate(data: any, id: string) {
+    const shallowClone = {...data}
+    delete shallowClone.id; // just incase left over from text area
+
+    const base = stripeServer + 'sources/'
+    const url = base + id;
+
+    request({
+      url,
+      post: shallowClone,
+      authorizationBearer: this.privateKey
+    }).then(result => {
+      this.source_update.result = tryParse(result)
+      this.source_update.resultAt = Date.now()
+    })
+  }
+
+  sendPaymentMethodUpdate(data: any, id: string) {
+    const shallowClone = {...data}
+    delete shallowClone.id; // just incase left over from text area
+
+    const base = stripeServer + 'payment_methods/'
+    const url = base + id;
+
+    request({
+      url,
+      post: shallowClone,
+      authorizationBearer: this.privateKey
+    }).then(result => {
+      this.payment_method_update.result = tryParse(result)
+      this.payment_method_update.resultAt = Date.now()
+    })
+  }
+
+  updateCustomer(data: any, id: string) {
+    const shallowClone = {...data}
+    delete shallowClone.id; // just incase left over from text area
+
+    const base = stripeServer + 'customers/'
+    const url = base + id;
+
+    request({
+      url,
+      post: shallowClone,
+      authorizationBearer: this.privateKey
+    }).then(result => {
+      this.customer_update.result = tryParse(result)
+      this.customer_update.resultAt = Date.now()
+    })
+  }
+
+  cleanSourceUpdateData(data: any) {
+    const deepClone = JSON.parse(JSON.stringify(data))
+
+    const removeKeys = [
+      'object', 'client_secret', 'created', 'flow', 'livemode', 'address', 'status', 'type', 'usage',
+    ]
+
+    this.cleanCardData(deepClone)
+    removeKeys.forEach(key => delete deepClone[key])
+    delete deepClone.owner.address
+
+    return deepClone
+  }
+
+  cleanCardData(data: any) {
+    const cardRemoveKeys = [
+      'three_d_secure_usage', 'fingerprint', 'last4', 'country', 'brand', 'address_line1_check', 'address_zip_check', 'cvc_check', 'funding', 'three_d_secure'
+    ]
+    cardRemoveKeys.forEach(key => delete data.card[key])
+    if (data.card.networks) {
+      delete data.card.networks.available
+    }
+  }
+
+  cleanPaymentMethodUpdateData(data: any) {
+    const deepClone = JSON.parse(JSON.stringify(data))
+
+    const removeKeys = [
+      'object', 'address', 'checks', 'available', 'created', 'livemode', 'type',
+    ]
+
+    removeKeys.forEach(key => delete deepClone[key])
+    this.cleanCardData(deepClone)
+    delete deepClone.billing_details.address
+
+    return deepClone
+  }
+
+  cleanCustomerUpdateData(data: any) {
+    const deepClone = JSON.parse(JSON.stringify(data))
+
+    const removeKeys = [
+      'account_balance', 'balance',
+      'object', 'cards', 'created', 'delinquent', 'livemode', 'sources', 'data', 'subscriptions', 'tax_ids'
+    ]
+    removeKeys.forEach(key => delete deepClone[key])
+
+    return deepClone
+  }
+}
 
 function request(
   {url, post, authorizationBearer}: {
@@ -298,7 +518,8 @@ function request(
 ) {
   return new Promise((res, rej) => {
     const req = new XMLHttpRequest();
-    req.open('POST', url, true);
+    const method = post ? 'POST' : 'GET'
+    req.open(method, url, true);
     req.setRequestHeader('Accept', 'application/json');
 
     if (authorizationBearer) {
@@ -371,5 +592,7 @@ function tryParse(data: string | any) {
 interface ISimpleRouteEditor {
   data: any
   result?: any
+  resultAt?: number
+  retrieve?: any // any update checks that might occur
   $send: EventEmitter<any>
 }